requirements.txt
Flask>=2.0
flask-session>=0.4
python-dotenv>=0.20
requests>=2.28
gunicorn>=20.1
ytmusicapi>=0.20
PyQt5>=5.15
PyQtWebEngine>=5.15
python-vlc>=3.0.20123
bcrypt>=4.0.0



